Thu, Dec 29, 2022, 7:30 PM +00:00

+++ project/lostark/expenses/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 # Create your models here.
 
 class Expense(models.Model):
-    amount = models.FloatField()
+    amount = models.IntegerField()
     date = models.DateField(default=now)
     description = models.TextField()
     owner = models.ForeignKey(to=User, on_delete=models.CASCADE)
