+++ project/lostark/income/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 # Create your models here.
 
 class Income(models.Model):
-    amount = models.FloatField()
+    amount = models.IntegerField()
     date = models.DateField(default=now)
     description = models.TextField()
     owner = models.ForeignKey(to=User, on_delete=models.CASCADE)
